+++ src/main/java/com/moment/the/util/controller/HttpRequestController.java
@@ -1,0 +1,2 @@
+package com.moment.the.util.controller;public class HttpRequestController {
+}
